More 64 fixes for semantic errors that only show up when templates are instantiated via unittests.  Now I'm actually serious about Phobos passing semantic analysis, except for a few modules due to weird druntime issues.

+++ std/container.d
@@ -314,7 +314,7 @@ Generally a container may define several types of ranges.
             assert(0);
         }
         /// Ditto
-        void opIndexAssign(T value, uint i)
+        void opIndexAssign(T value, size_t i)
         {
             assert(0);
         }
@@ -2775,7 +2775,7 @@ struct Array(T) if (is(T == bool))
             return _outer[_a + i];
         }
         /// Ditto
-        void opIndexAssign(T value, uint i)
+        void opIndexAssign(T value, size_t i)
         {
             _outer[_a + i] = value;
         }
